@@ -1,6 +1,6 @@
 {
   "name": "followjs",
-  "version": "1.1.8",
+  "version": "1.1.10",
   "productName": "FollowJS",
   "description": "ArtNet FollowSpot control",
   "main": "index.js",
@@ -14,7 +14,7 @@
   },
   "private": true,
   "devDependencies": {
-    "electron": "^13.1.9",
+    "electron": "^16.0.5",
     "electron-builder": "^22.13.1",
     "electron-rebuild": "^3.1.1"
   },
