@@ -79,9 +79,9 @@ module.exports = {
                 15: {name: 'M.Frost', short: 'MF', value: 0},
                 16: {name: 'H.Frost', short: 'HF', value: 0},
                 17: {name: 'Focus', short: 'F', value: 127},
-                18: {name: 'Pan', short: 'P', value: 0},
+                18: {name: 'Pan', short: 'Pan', value: 0},
                 19: {name: 'PanFine', short: 'P16', value: 0},
-                20: {name: 'Tilt', short: 'T', value: 0},
+                20: {name: 'Tilt', short: 'Tilt', value: 0},
                 21: {name: 'TiltFine',short: 'T16', value: 0},
                 22: {name: 'Function', short: 'Func', value: 0},
                 23: {name: 'Reset', short: 'Res', value: 0},
